+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/gifts/GetChatGifts.kt
@@ -4,7 +4,7 @@ import dev.inmo.tgbotapi.bot.TelegramBot
 import dev.inmo.tgbotapi.requests.gifts.GetChatGifts
 import dev.inmo.tgbotapi.types.ChatIdentifier
 import dev.inmo.tgbotapi.types.OwnedGifts
-import dev.inmo.tgbotapi.types.gifts.GiftSentOrReceived
+import dev.inmo.tgbotapi.types.gifts.GiftSentOrReceivedEvent
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
@@ -20,7 +20,7 @@ public suspend fun TelegramBot.getChatGifts(
     sortByPrice: Boolean = false,
     offset: String? = null,
     limit: Int? = null
-): OwnedGifts<GiftSentOrReceived> = execute(
+): OwnedGifts<GiftSentOrReceivedEvent> = execute(
     GetChatGifts(
         chatId,
         excludeUnsaved,
@@ -49,7 +49,7 @@ public fun TelegramBot.getChatGiftsFlow(
     initialOffset: String? = null,
     limit: Int? = null,
     onErrorContinueChecker: suspend (Throwable?) -> Boolean = { false }
-): Flow<OwnedGifts<GiftSentOrReceived>> = flow {
+): Flow<OwnedGifts<GiftSentOrReceivedEvent>> = flow {
     var currentOffset = initialOffset
     do {
         val response = runCatching {