+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `Core`:
     * **THIS IS BREAKING CHANGE** All media files with sizes changed type of `fileSize` field from `Long` to `FileSize` (value class)
+    * **THIS IS BREAKING CHANGE** Class `GiftSentOrReceived` have been renamed to `GiftSentOrReceivedEvent` to clarify naming
     * **THIS IS BREAKING CHANGE** For `CheclistTask` have been made several changes:
       * `ChecklistTask.Done` -> `ChecklistTask.Completed`
       * `ChecklistTask.Undone` -> `ChecklistTask.Uncompleted`